+++ src/main/java/org/cryptomator/common/integrations/sslcontext/SSLContextWithMacKeychain.java
@@ -2,14 +2,10 @@ package org.cryptomator.common.integrations.sslcontext;
 
 import org.cryptomator.integrations.common.OperatingSystem;
 
-import javax.net.ssl.SSLContext;
-import javax.net.ssl.TrustManagerFactory;
 import java.io.IOException;
-import java.security.KeyManagementException;
 import java.security.KeyStore;
 import java.security.KeyStoreException;
 import java.security.NoSuchAlgorithmException;
-import java.security.SecureRandom;
 import java.security.cert.CertificateException;
 
 /**
@@ -18,22 +14,10 @@ import java.security.cert.CertificateException;
  * Provided by java.base module/jmod
  */
 @OperatingSystem(OperatingSystem.Value.MAC)
-public class SSLContextWithMacKeychain implements SSLContextProvider {
+public class SSLContextWithMacKeychain extends SSLContextDifferentTrustStoreBase implements SSLContextProvider {
 
 	@Override
-	public SSLContext getContext(SecureRandom csprng) throws SSLContextBuildException {
-		try {
-			KeyStore truststore = KeyStore.getInstance("KeychainStore-ROOT");
-			truststore.load(null, null);
-
-			TrustManagerFactory tmf = TrustManagerFactory.getInstance(TrustManagerFactory.getDefaultAlgorithm());
-			tmf.init(truststore);
-
-			SSLContext context = SSLContext.getInstance("TLS");
-			context.init(null, tmf.getTrustManagers(), csprng);
-			return context;
-		} catch (KeyStoreException | CertificateException | NoSuchAlgorithmException | KeyManagementException | IOException e) {
-			throw new SSLContextBuildException(e);
-		}
+	KeyStore getTruststore() throws KeyStoreException, CertificateException, IOException, NoSuchAlgorithmException {
+		return KeyStore.getInstance("KeychainStore-ROOT");
 	}
 }
