+++ src/main/java/org/cryptomator/common/locationpresets/LocationPresetsProvider.java
@@ -1,18 +1,10 @@
 package org.cryptomator.common.locationpresets;
 
 import org.cryptomator.common.integrations.IntegrationsLoaderCopy;
-import org.cryptomator.integrations.common.CheckAvailability;
-import org.cryptomator.integrations.common.IntegrationsLoader;
-import org.cryptomator.integrations.common.OperatingSystem;
-import org.jetbrains.annotations.Nullable;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.lang.reflect.Method;
-import java.lang.reflect.Modifier;
 import java.nio.file.Path;
-import java.util.Arrays;
-import java.util.ServiceLoader;
 import java.util.stream.Stream;
 
 public interface LocationPresetsProvider {
@@ -34,8 +26,6 @@ public interface LocationPresetsProvider {
 		}
 	}
 
-	//copied from org.cryptomator.integrations.common.IntegrationsLoader
-	//TODO: delete, once migrated to integrations-api
 	static Stream<LocationPresetsProvider> loadAll() {
 		return IntegrationsLoaderCopy.loadAll(LocationPresetsProvider.class);
 	}
